refactor(positions): colocate integration tests next to transform module
