+++ src/features/positions/stores/transform/compound.test.ts
@@ -1,8 +1,8 @@
-import { transformPositions } from '../../stores/transform';
+import { transformPositions } from '.';
 import { PositionName, DetailType } from '../../types/generated/positions/positions';
 import { FIXTURE_PARAMS } from '../../__fixtures__/ListPositions';
-import { createMockAsset } from '../mocks/assets';
-import { createMockStats, createMockPosition, createMockResponse } from '../mocks/positions';
+import { createMockAsset } from '../../__fixtures__/mocks/assets';
+import { createMockStats, createMockPosition, createMockResponse } from '../../__fixtures__/mocks/positions';
 
 // Mock config to avoid React Native gesture handler imports
 jest.mock('@/config', () => ({
